requirements.txt
attrs>=24.3.0
blinker>=1.9.0
certifi>=2024.12.14
click>=8.1.8
exceptiongroup>=1.2.2
Flask>=3.1.0
h11>=0.14.0
idna>=3.10
importlib_metadata>=8.5.0
itsdangerous>=2.2.0
Jinja2>=3.1.5
MarkupSafe>=3.0.2
outcome>=1.3.0.post0
PySocks>=1.7.1
PyYAML>=6.0.2
selenium>=4.27.1
sniffio>=1.3.1
sortedcontainers>=2.4.0
trio>=0.28.0
trio-websocket>=0.11.1
typing_extensions>=4.12.2
urllib3>=2.3.0
websocket-client>=1.8.0
Werkzeug>=3.1.3
wsproto>=1.2.0
zipp>=3.21.0
